Revert "Extract __local endpoints into shared module for Vite + Bun.serve (#32756)"
This reverts commit 79d38b3902000a9a53cc1449f55dbb90d6bddc87.
