+++ apps/web/vite-plugin-local-mode.ts
@@ -1,54 +1,141 @@
-import http from "node:http";
+import { spawn } from "node:child_process";
 import fs from "node:fs";
+import http from "node:http";
+import { createRequire } from "node:module";
+import os from "node:os";
 import path from "node:path";
 import type { Plugin, Connect, ViteDevServer } from "vite";
+import { SEEDS } from "../../cli/src/lib/environments/seeds";
 
-import {
-  resolveLocalConfigFromEnv,
-  resolveCliPath,
-  isLoopbackAddr,
-  getLockfileData,
-  upsertLockfileAssistant,
-  runHatch,
-  runRetire,
-  getGuardianAccessToken,
-  parseGatewayUrl,
-  readAllowedGatewayPorts,
-} from "../../cli/src/lib/local-endpoints";
+const PRODUCTION_ENVIRONMENT_NAME = "production";
+const CLI_PACKAGE_NAME = "@vellumai/cli";
+const LOCALHOST_ORIGIN_RE = /^https?:\/\/(localhost|127\.0\.0\.1)(:\d+)?$/;
 
-const GUARDIAN_TOKEN_PATTERN =
-  /^(?:\/assistant)?\/__local\/guardian-token\/([^/]+)$/;
+let _resolvedCliPath: string | undefined;
 
+/**
+ * Resolve the CLI entry point via two strategies:
+ *
+ * 1. **Source tree** — `<repoRoot>/cli/src/index.ts` exists (dev mode in monorepo).
+ * 2. **Installed package** — `require.resolve("@vellumai/cli/package.json")` then
+ *    derive the entry point from the resolved package directory.
+ *
+ * The result is cached for the lifetime of the Vite server process.
+ */
+function resolveCliPath(): string {
+  if (_resolvedCliPath) return _resolvedCliPath;
+
+  const repoRoot = path.resolve(import.meta.dirname, "..", "..");
+  const sourceTreePath = path.join(repoRoot, "cli", "src", "index.ts");
+  if (fs.existsSync(sourceTreePath)) {
+    _resolvedCliPath = sourceTreePath;
+    return _resolvedCliPath;
+  }
+
+  const _require = createRequire(import.meta.url);
+  try {
+    const pkgPath = _require.resolve(`${CLI_PACKAGE_NAME}/package.json`);
+    const pkg = JSON.parse(fs.readFileSync(pkgPath, "utf-8")) as { bin?: Record<string, string> };
+    const binEntry = pkg.bin?.["vellum"];
+    if (binEntry) {
+      const entryPoint = path.resolve(path.dirname(pkgPath), binEntry);
+      if (fs.existsSync(entryPoint)) {
+        _resolvedCliPath = entryPoint;
+        return _resolvedCliPath;
+      }
+    }
+  } catch {
+    // Not found in node_modules
+  }
+
+  throw new Error(
+    `Vellum CLI not found. Looked for source tree at ${sourceTreePath} and npm package ${CLI_PACKAGE_NAME}.`,
+  );
+}
+
+/**
+ * Sensitive lockfile fields that must never be served to the browser.
+ */
+const SENSITIVE_FIELDS = [
+  "signingKey",
+  "bearerToken",
+  "guardianBootstrapSecret",
+] as const;
+
+/**
+ * Resolve the lockfile path on disk using XDG conventions.
+ *
+ * Mirrors `cli/src/lib/environments/paths.ts` logic:
+ * - Production: `~/.vellum.lock.json`
+ * - Non-production: `$XDG_CONFIG_HOME/vellum-{env}/lockfile.json`
+ * - `VELLUM_LOCKFILE_DIR` overrides the directory in both cases.
+ */
+function resolveLockfilePaths(env: Record<string, string>): string[] {
+  const vellumEnv = env.VELLUM_ENVIRONMENT || PRODUCTION_ENVIRONMENT_NAME;
+  const lockfileDir = env.VELLUM_LOCKFILE_DIR;
+
+  if (vellumEnv === PRODUCTION_ENVIRONMENT_NAME) {
+    const dir = lockfileDir ?? os.homedir();
+    return [
+      path.join(dir, ".vellum.lock.json"),
+      path.join(dir, ".vellum.lockfile.json"),
+    ];
+  }
+
+  const xdgConfigHome =
+    env.XDG_CONFIG_HOME || path.join(os.homedir(), ".config");
+  const dir = lockfileDir ?? path.join(xdgConfigHome, `vellum-${vellumEnv}`);
+  return [path.join(dir, "lockfile.json")];
+}
+
+/**
+ * Strip sensitive fields from each assistant entry in the lockfile data.
+ */
+function stripSensitiveFields(data: Record<string, unknown>): void {
+  const assistants = data.assistants;
+  if (!Array.isArray(assistants)) return;
+  for (const assistant of assistants) {
+    if (assistant && typeof assistant === "object") {
+      const entry = assistant as Record<string, unknown>;
+      for (const field of SENSITIVE_FIELDS) {
+        delete entry[field];
+      }
+      const resources = entry.resources;
+      if (resources && typeof resources === "object") {
+        for (const field of SENSITIVE_FIELDS) {
+          delete (resources as Record<string, unknown>)[field];
+        }
+      }
+    }
+  }
+}
+
+/**
+ * Vite plugin that serves lockfile endpoints and a dynamic gateway proxy
+ * for local-mode development.
+ */
 export function localModePlugin(env: Record<string, string>): Plugin {
-  const config = resolveLocalConfigFromEnv(env);
-  const baseDir = path.resolve(import.meta.dirname, "..", "..");
+  const lockfilePaths = resolveLockfilePaths(env);
 
   return {
     name: "vellum-local-mode",
     configureServer(server) {
       server.middlewares.use(loopbackCallbackMiddleware());
-      server.middlewares.use(configMiddleware(config.webUrl, config.platformUrl));
-      server.middlewares.use(lockfileMiddleware(config.lockfilePaths));
-      server.middlewares.use(hatchMiddleware(baseDir));
-      server.middlewares.use(retireMiddleware(baseDir));
-      server.middlewares.use(guardianTokenMiddleware(config.configDir, baseDir, env));
-      server.middlewares.use(gatewayProxyMiddleware(config.lockfilePaths));
+      server.middlewares.use(configMiddleware(env));
+      server.middlewares.use(lockfileMiddleware(lockfilePaths));
+      server.middlewares.use(hatchMiddleware());
+      server.middlewares.use(retireMiddleware(lockfilePaths));
+      server.middlewares.use(guardianTokenMiddleware(env));
+      server.middlewares.use(gatewayProxyMiddleware(lockfilePaths));
       server.middlewares.use(accountSpaFallback(server));
     },
   };
 }
 
-function rejectUnlessLoopback(
-  req: http.IncomingMessage,
-  res: http.ServerResponse,
-): boolean {
-  if (isLoopbackAddr(req.socket.remoteAddress ?? "")) return false;
-  res.statusCode = 403;
-  res.setHeader("Content-Type", "application/json");
-  res.end(JSON.stringify({ error: "Forbidden" }));
-  return true;
-}
-
+/**
+ * Redirect `/callback` from the platform's CLI loopback auth flow
+ * into the SPA route that processes the session token.
+ */
 function loopbackCallbackMiddleware(): Connect.NextHandleFunction {
   return (req, res, next) => {
     if (req.url?.startsWith("/callback")) {
@@ -61,7 +148,16 @@ function loopbackCallbackMiddleware(): Connect.NextHandleFunction {
   };
 }
 
-function configMiddleware(webUrl: string, platformUrl: string): Connect.NextHandleFunction {
+/**
+ * Serve environment config so the SPA can resolve the platform web URL
+ * for loopback auth without hardcoding it.
+ */
+function configMiddleware(env: Record<string, string>): Connect.NextHandleFunction {
+  const vellumEnv = env.VELLUM_ENVIRONMENT || PRODUCTION_ENVIRONMENT_NAME;
+  const seed = SEEDS[vellumEnv] ?? SEEDS[PRODUCTION_ENVIRONMENT_NAME]!;
+
+  const webUrl = env.VELLUM_WEB_URL || seed.webUrl;
+  const platformUrl = env.VELLUM_PLATFORM_URL || seed.platformUrl;
   const body = JSON.stringify({ webUrl, platformUrl });
 
   return (req, res, next) => {
@@ -71,6 +167,12 @@ function configMiddleware(webUrl: string, platformUrl: string): Connect.NextHand
   };
 }
 
+/**
+ * SPA fallback for `/account/*` routes. Vite's base is `/assistant/` so
+ * paths outside it don't get the default SPA fallback. This middleware
+ * serves the transformed index.html for account routes so React Router
+ * can handle them.
+ */
 function accountSpaFallback(server: ViteDevServer): Connect.NextHandleFunction {
   return (req, res, next) => {
     if (!req.url?.startsWith("/account/") && !req.url?.startsWith("/account?") && req.url !== "/account") return next();
@@ -89,44 +191,21 @@ function accountSpaFallback(server: ViteDevServer): Connect.NextHandleFunction {
   };
 }
 
-function lockfileMiddleware(lockfilePaths: string[]): Connect.NextHandleFunction {
+/**
+ * Connect middleware for the lockfile read endpoint.
+ */
+function lockfileMiddleware(
+  lockfilePaths: string[],
+): Connect.NextHandleFunction {
   return (req, res, next) => {
     if (req.url !== "/assistant/__local/lockfile" && req.url !== "/__local/lockfile") return next();
 
     if (rejectUnlessLoopback(req, res)) return;
 
     if (req.method === "GET") {
-      const result = getLockfileData(lockfilePaths);
-      if (result.ok) {
-        res.setHeader("Content-Type", "application/json");
-        res.end(JSON.stringify(result.data));
-      } else {
-        res.statusCode = result.status;
-        res.end();
-      }
+      handleGetLockfile(lockfilePaths, res);
     } else if (req.method === "POST") {
-      const chunks: Buffer[] = [];
-      req.on("data", (chunk: Buffer) => chunks.push(chunk));
-      req.on("end", () => {
-        let body: Record<string, unknown>;
-        try {
-          body = JSON.parse(Buffer.concat(chunks).toString()) as Record<string, unknown>;
-        } catch {
-          res.statusCode = 400;
-          res.setHeader("Content-Type", "application/json");
-          res.end(JSON.stringify({ ok: false, error: "Invalid JSON body" }));
-          return;
-        }
-
-        const result = upsertLockfileAssistant(
-          lockfilePaths,
-          body.assistant as Record<string, unknown>,
-          body.activeAssistant as string | undefined,
-        );
-        res.statusCode = result.ok ? 200 : result.status;
-        res.setHeader("Content-Type", "application/json");
-        res.end(JSON.stringify(result));
-      });
+      handlePostLockfile(lockfilePaths, req, res);
     } else {
       res.statusCode = 405;
       res.end();
@@ -134,9 +213,139 @@ function lockfileMiddleware(lockfilePaths: string[]): Connect.NextHandleFunction
   };
 }
 
-function hatchMiddleware(baseDir: string): Connect.NextHandleFunction {
+function handleGetLockfile(
+  lockfilePaths: string[],
+  res: http.ServerResponse,
+): void {
+  let raw: string | undefined;
+  for (const candidate of lockfilePaths) {
+    try {
+      raw = fs.readFileSync(candidate, "utf-8");
+      break;
+    } catch (err: unknown) {
+      if ((err as NodeJS.ErrnoException).code !== "ENOENT") {
+        res.statusCode = 500;
+        res.end();
+        return;
+      }
+    }
+  }
+
+  if (!raw) {
+    res.setHeader("Content-Type", "application/json");
+    res.end(JSON.stringify({ assistants: [], activeAssistant: null }));
+    return;
+  }
+
+  try {
+    const data = JSON.parse(raw) as Record<string, unknown>;
+    stripSensitiveFields(data);
+    res.setHeader("Content-Type", "application/json");
+    res.end(JSON.stringify(data));
+  } catch {
+    res.statusCode = 500;
+    res.end();
+  }
+}
+
+/**
+ * Merge an assistant entry into the lockfile on disk.
+ *
+ * Transport: Vite dev middleware (fs read/write).
+ * In Electron, replace with IPC call to main process: window.electronAPI.saveLockfileAssistant(entry). (LUM-1998)
+ */
+function handlePostLockfile(
+  lockfilePaths: string[],
+  req: http.IncomingMessage,
+  res: http.ServerResponse,
+): void {
+  const chunks: Buffer[] = [];
+  req.on("data", (chunk: Buffer) => chunks.push(chunk));
+  req.on("end", () => {
+    let body: Record<string, unknown>;
+    try {
+      body = JSON.parse(Buffer.concat(chunks).toString()) as Record<string, unknown>;
+    } catch {
+      res.statusCode = 400;
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify({ ok: false, error: "Invalid JSON body" }));
+      return;
+    }
+
+    const assistant = body.assistant as Record<string, unknown> | undefined;
+    const activeAssistant = body.activeAssistant as string | undefined;
+    if (!assistant || typeof assistant.assistantId !== "string") {
+      res.statusCode = 400;
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify({ ok: false, error: "Missing assistant.assistantId" }));
+      return;
+    }
+
+    // Read existing lockfile
+    let lockfile: Record<string, unknown> = { assistants: [], activeAssistant: null };
+    const writePath = lockfilePaths[0]!;
+    for (const candidate of lockfilePaths) {
+      try {
+        lockfile = JSON.parse(fs.readFileSync(candidate, "utf-8")) as Record<string, unknown>;
+        break;
+      } catch {
+        // continue
+      }
+    }
+
+    // Upsert the assistant entry
+    const assistants = Array.isArray(lockfile.assistants) ? lockfile.assistants : [];
+    const existingIdx = assistants.findIndex(
+      (a: Record<string, unknown>) => a?.assistantId === assistant.assistantId,
+    );
+    if (existingIdx >= 0) {
+      assistants[existingIdx] = { ...assistants[existingIdx], ...assistant };
+    } else {
+      assistants.push(assistant);
+    }
+    lockfile.assistants = assistants;
+    if (activeAssistant !== undefined) {
+      lockfile.activeAssistant = activeAssistant;
+    }
+
+    // Atomic write
+    try {
+      const dir = path.dirname(writePath);
+      fs.mkdirSync(dir, { recursive: true });
+      const tmp = `${writePath}.tmp.${process.pid}`;
+      fs.writeFileSync(tmp, JSON.stringify(lockfile, null, 2));
+      fs.renameSync(tmp, writePath);
+    } catch (err) {
+      res.statusCode = 500;
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify({ ok: false, error: `Failed to write lockfile: ${err}` }));
+      return;
+    }
+
+    const stripped = JSON.parse(JSON.stringify(lockfile)) as Record<string, unknown>;
+    stripSensitiveFields(stripped);
+    res.setHeader("Content-Type", "application/json");
+    res.end(JSON.stringify({ ok: true, lockfile: stripped }));
+  });
+}
+
+const HATCH_TIMEOUT_MS = 120_000;
+
+/**
+ * Connect middleware for the hatch endpoint.
+ *
+ * Transport: Vite dev middleware (child_process.spawn → CLI binary).
+ * In Electron, replace with IPC call to main process: window.electronAPI.hatchAssistant(species). (LUM-1997)
+ * The main process has direct access to the hatch-local module without spawning a subprocess.
+ */
+function hatchMiddleware(): Connect.NextHandleFunction {
   return (req, res, next) => {
-    if (req.url !== "/assistant/__local/hatch" && req.url !== "/__local/hatch") return next();
+    if (
+      req.url !== "/assistant/__local/hatch" &&
+      req.url !== "/__local/hatch"
+    ) {
+      return next();
+    }
 
     if (rejectUnlessLoopback(req, res)) return;
 
@@ -152,8 +361,12 @@ function hatchMiddleware(baseDir: string): Connect.NextHandleFunction {
       let species = "vellum";
       if (chunks.length > 0) {
         try {
-          const body = JSON.parse(Buffer.concat(chunks).toString()) as { species?: string };
-          if (body.species) species = body.species;
+          const body = JSON.parse(Buffer.concat(chunks).toString()) as {
+            species?: string;
+          };
+          if (body.species) {
+            species = body.species;
+          }
         } catch {
           res.statusCode = 400;
           res.setHeader("Content-Type", "application/json");
@@ -162,28 +375,86 @@ function hatchMiddleware(baseDir: string): Connect.NextHandleFunction {
         }
       }
 
-      let cliPath: string;
-      try {
-        cliPath = resolveCliPath(baseDir, import.meta.url);
-      } catch (err) {
-        res.statusCode = 500;
-        res.setHeader("Content-Type", "application/json");
-        res.end(JSON.stringify({ ok: false, error: err instanceof Error ? err.message : String(err) }));
-        return;
-      }
-
-      runHatch(species, cliPath).then((result) => {
-        res.statusCode = result.ok ? 200 : result.status;
-        res.setHeader("Content-Type", "application/json");
-        res.end(JSON.stringify(result.ok ? { ok: true, assistantId: result.assistantId } : { ok: false, error: result.error }));
-      });
+      handleHatch(species, res);
     });
   };
 }
 
-function retireMiddleware(baseDir: string): Connect.NextHandleFunction {
+function handleHatch(species: string, res: http.ServerResponse): void {
+  let cliPath: string;
+  try {
+    cliPath = resolveCliPath();
+  } catch (err) {
+    res.statusCode = 500;
+    res.setHeader("Content-Type", "application/json");
+    res.end(
+      JSON.stringify({
+        ok: false,
+        error: err instanceof Error ? err.message : String(err),
+      }),
+    );
+    return;
+  }
+
+  const child = spawn("bun", ["run", cliPath, "hatch", species], {
+    stdio: ["ignore", "pipe", "pipe"],
+  });
+
+  let stdout = "";
+  let stderr = "";
+  let responded = false;
+
+  const respond = (status: number, body: Record<string, unknown>) => {
+    if (responded) return;
+    responded = true;
+    clearTimeout(timeout);
+    res.statusCode = status;
+    res.setHeader("Content-Type", "application/json");
+    res.end(JSON.stringify(body));
+  };
+
+  const timeout = setTimeout(() => {
+    child.kill("SIGTERM");
+    respond(500, { ok: false, error: "Hatch timed out after 120 seconds" });
+  }, HATCH_TIMEOUT_MS);
+
+  child.stdout.on("data", (data: Buffer) => {
+    stdout += data.toString();
+  });
+
+  child.stderr.on("data", (data: Buffer) => {
+    stderr += data.toString();
+  });
+
+  child.on("close", (code) => {
+    if (code === 0) {
+      const match = stdout.match(/Hatching local assistant:\s+(.+)/);
+      const assistantId = match?.[1]?.trim() ?? "";
+      respond(200, { ok: true, assistantId });
+    } else {
+      respond(500, { ok: false, error: stderr || stdout });
+    }
+  });
+
+  child.on("error", (err) => {
+    respond(500, { ok: false, error: `Failed to spawn CLI: ${err.message}` });
+  });
+}
+
+/**
+ * Connect middleware for the retire endpoint.
+ *
+ * Transport: Vite dev middleware (child_process.spawn → CLI binary).
+ * In Electron, replace with IPC call to main process: window.electronAPI.retireAssistant(assistantId). (LUM-2000)
+ */
+function retireMiddleware(lockfilePaths: string[]): Connect.NextHandleFunction {
   return (req, res, next) => {
-    if (req.url !== "/assistant/__local/retire" && req.url !== "/__local/retire") return next();
+    if (
+      req.url !== "/assistant/__local/retire" &&
+      req.url !== "/__local/retire"
+    ) {
+      return next();
+    }
 
     if (rejectUnlessLoopback(req, res)) return;
 
@@ -193,13 +464,31 @@ function retireMiddleware(baseDir: string): Connect.NextHandleFunction {
       return;
     }
 
+    const peer = req.socket.remoteAddress ?? "";
+    if (!isLoopbackAddr(peer)) {
+      res.statusCode = 403;
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify({ ok: false, error: "Forbidden" }));
+      return;
+    }
+
+    const origin = req.headers.origin;
+    if (!origin || !LOCALHOST_ORIGIN_RE.test(origin)) {
+      res.statusCode = 403;
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify({ ok: false, error: "Forbidden" }));
+      return;
+    }
+
     const chunks: Buffer[] = [];
     req.on("data", (chunk: Buffer) => chunks.push(chunk));
     req.on("end", () => {
       let assistantId: string | undefined;
       if (chunks.length > 0) {
         try {
-          const body = JSON.parse(Buffer.concat(chunks).toString()) as { assistantId?: string };
+          const body = JSON.parse(Buffer.concat(chunks).toString()) as {
+            assistantId?: string;
+          };
           assistantId = body.assistantId;
         } catch {
           res.statusCode = 400;
@@ -216,28 +505,188 @@ function retireMiddleware(baseDir: string): Connect.NextHandleFunction {
         return;
       }
 
-      let cliPath: string;
-      try {
-        cliPath = resolveCliPath(baseDir, import.meta.url);
-      } catch (err) {
-        res.statusCode = 500;
+      const activeId = getActiveLocalAssistantId(lockfilePaths);
+      if (!activeId || assistantId !== activeId) {
+        res.statusCode = 403;
         res.setHeader("Content-Type", "application/json");
-        res.end(JSON.stringify({ ok: false, error: err instanceof Error ? err.message : String(err) }));
+        res.end(JSON.stringify({ ok: false, error: "Forbidden" }));
         return;
       }
 
-      runRetire(assistantId, cliPath).then((result) => {
-        res.statusCode = result.ok ? 200 : result.status;
-        res.setHeader("Content-Type", "application/json");
-        res.end(JSON.stringify(result.ok ? { ok: true } : { ok: false, error: result.error }));
-      });
+      handleRetire(assistantId, res);
     });
   };
 }
 
+function getActiveLocalAssistantId(lockfilePaths: string[]): string | undefined {
+  for (const candidate of lockfilePaths) {
+    try {
+      const raw = fs.readFileSync(candidate, "utf-8");
+      const parsed = JSON.parse(raw) as {
+        activeAssistant?: unknown;
+        assistants?: Array<Record<string, unknown>>;
+      };
+
+      if (typeof parsed.activeAssistant !== "string") {
+        return undefined;
+      }
+
+      const activeEntry = parsed.assistants?.find(
+        (a) => a?.assistantId === parsed.activeAssistant,
+      );
+      if (!activeEntry || activeEntry.cloud !== "local") {
+        return undefined;
+      }
+
+      return parsed.activeAssistant;
+    } catch (err: unknown) {
+      if ((err as NodeJS.ErrnoException).code === "ENOENT") {
+        continue;
+      }
+      return undefined;
+    }
+  }
+  return undefined;
+}
+
+const RETIRE_TIMEOUT_MS = 60_000;
+
+function handleRetire(assistantId: string, res: http.ServerResponse): void {
+  let cliPath: string;
+  try {
+    cliPath = resolveCliPath();
+  } catch (err) {
+    res.statusCode = 500;
+    res.setHeader("Content-Type", "application/json");
+    res.end(
+      JSON.stringify({
+        ok: false,
+        error: err instanceof Error ? err.message : String(err),
+      }),
+    );
+    return;
+  }
+
+  const child = spawn("bun", ["run", cliPath, "retire", assistantId, "--yes"], {
+    stdio: ["ignore", "pipe", "pipe"],
+  });
+
+  let stdout = "";
+  let stderr = "";
+  let responded = false;
+
+  const respond = (status: number, body: Record<string, unknown>) => {
+    if (responded) return;
+    responded = true;
+    clearTimeout(timeout);
+    res.statusCode = status;
+    res.setHeader("Content-Type", "application/json");
+    res.end(JSON.stringify(body));
+  };
+
+  const timeout = setTimeout(() => {
+    child.kill("SIGTERM");
+    respond(500, { ok: false, error: "Retire timed out after 60 seconds" });
+  }, RETIRE_TIMEOUT_MS);
+
+  child.stdout.on("data", (data: Buffer) => {
+    stdout += data.toString();
+  });
+
+  child.stderr.on("data", (data: Buffer) => {
+    stderr += data.toString();
+  });
+
+  child.on("close", (code) => {
+    if (code === 0) {
+      respond(200, { ok: true });
+    } else {
+      respond(500, { ok: false, error: stderr || stdout });
+    }
+  });
+
+  child.on("error", (err) => {
+    respond(500, { ok: false, error: `Failed to spawn CLI: ${err.message}` });
+  });
+}
+
+// ---------------------------------------------------------------------------
+// Loopback guard
+// ---------------------------------------------------------------------------
+
+function isLoopbackAddr(addr: string): boolean {
+  const v4Mapped = addr.match(/^::ffff:(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})$/i);
+  const normalized = v4Mapped ? v4Mapped[1]! : addr;
+  if (normalized.includes(".")) {
+    return normalized.startsWith("127.");
+  }
+  return normalized === "::1";
+}
+
+function rejectUnlessLoopback(
+  req: http.IncomingMessage,
+  res: http.ServerResponse,
+): boolean {
+  if (isLoopbackAddr(req.socket.remoteAddress ?? "")) return false;
+  res.statusCode = 403;
+  res.setHeader("Content-Type", "application/json");
+  res.end(JSON.stringify({ error: "Forbidden" }));
+  return true;
+}
+
+const GUARDIAN_TOKEN_PATTERN =
+  /^(?:\/assistant)?\/__local\/guardian-token\/([^/]+)$/;
+
+const GUARDIAN_TOKEN_REFRESH_TIMEOUT_MS = 15_000;
+
+/**
+ * Resolve the config directory matching `cli/src/lib/environments/paths.ts:getConfigDir`.
+ */
+function resolveConfigDir(env: Record<string, string>): string {
+  const vellumEnv = env.VELLUM_ENVIRONMENT || PRODUCTION_ENVIRONMENT_NAME;
+  const xdgConfigHome =
+    env.XDG_CONFIG_HOME || path.join(os.homedir(), ".config");
+  if (vellumEnv === PRODUCTION_ENVIRONMENT_NAME) {
+    return path.join(xdgConfigHome, "vellum");
+  }
+  return path.join(xdgConfigHome, `vellum-${vellumEnv}`);
+}
+
+function resolveGuardianTokenPath(
+  env: Record<string, string>,
+  assistantId: string,
+): string {
+  return path.join(resolveConfigDir(env), "assistants", assistantId, "guardian-token.json");
+}
+
+interface GuardianTokenData {
+  accessToken: string;
+  accessTokenExpiresAt: string | number;
+  refreshToken: string;
+  refreshTokenExpiresAt: string | number;
+}
+
+function isAccessTokenExpired(data: GuardianTokenData): boolean {
+  const expiresAt = new Date(data.accessTokenExpiresAt).getTime();
+  if (!Number.isFinite(expiresAt)) return true;
+  return Date.now() >= expiresAt - 60_000;
+}
+
+function isRefreshTokenExpired(data: GuardianTokenData): boolean {
+  const expiresAt = new Date(data.refreshTokenExpiresAt).getTime();
+  if (!Number.isFinite(expiresAt)) return true;
+  return Date.now() >= expiresAt;
+}
+
+/**
+ * Connect middleware that serves guardian access tokens for local assistants.
+ *
+ * GET /assistant/__local/guardian-token/:assistantId
+ *
+ * Reads the guardian token from disk. If the access token is expired,
+ * shells out to the CLI to refresh it, then re-reads from disk.
+ */
 function guardianTokenMiddleware(
-  configDir: string,
-  baseDir: string,
   env: Record<string, string>,
 ): Connect.NextHandleFunction {
   return (req, res, next) => {
@@ -253,10 +702,45 @@ function guardianTokenMiddleware(
     if (rejectUnlessLoopback(req, res)) return;
 
     const assistantId = decodeURIComponent(match[1]!);
+    const tokenPath = resolveGuardianTokenPath(env, assistantId);
 
+    let raw: string;
+    try {
+      raw = fs.readFileSync(tokenPath, "utf-8");
+    } catch {
+      res.statusCode = 404;
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify({ error: "Guardian token not found" }));
+      return;
+    }
+
+    let data: GuardianTokenData;
+    try {
+      data = JSON.parse(raw) as GuardianTokenData;
+    } catch {
+      res.statusCode = 500;
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify({ error: "Malformed guardian token file" }));
+      return;
+    }
+
+    if (!isAccessTokenExpired(data)) {
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify({ accessToken: data.accessToken }));
+      return;
+    }
+
+    if (isRefreshTokenExpired(data)) {
+      res.statusCode = 401;
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify({ error: "Guardian token expired — re-run `vellum hatch` or `vellum wake`" }));
+      return;
+    }
+
+    // Refresh via CLI in a child process
     let cliPath: string;
     try {
-      cliPath = resolveCliPath(baseDir, import.meta.url);
+      cliPath = resolveCliPath();
     } catch (err) {
       res.statusCode = 500;
       res.setHeader("Content-Type", "application/json");
@@ -264,46 +748,137 @@ function guardianTokenMiddleware(
       return;
     }
 
-    getGuardianAccessToken(assistantId, configDir, cliPath, true, env).then((result) => {
-      if (result.ok) {
-        res.setHeader("Content-Type", "application/json");
-        res.end(JSON.stringify({ accessToken: result.accessToken }));
+    const child = spawn(
+      "bun",
+      ["run", cliPath, "gateway", "token", "refresh", assistantId],
+      { stdio: ["ignore", "pipe", "pipe"], env: { ...process.env, ...env } },
+    );
+
+    let stdout = "";
+    let responded = false;
+
+    const respond = (status: number, body: Record<string, unknown>) => {
+      if (responded) return;
+      responded = true;
+      clearTimeout(timeout);
+      res.statusCode = status;
+      res.setHeader("Content-Type", "application/json");
+      res.end(JSON.stringify(body));
+    };
+
+    const timeout = setTimeout(() => {
+      child.kill("SIGTERM");
+      respond(500, { error: "Guardian token refresh timed out" });
+    }, GUARDIAN_TOKEN_REFRESH_TIMEOUT_MS);
+
+    child.stdout.on("data", (chunk: Buffer) => {
+      stdout += chunk.toString();
+    });
+
+    child.on("close", (code) => {
+      if (code === 0) {
+        const accessToken = stdout.trim();
+        if (accessToken) {
+          respond(200, { accessToken });
+        } else {
+          respond(500, { error: "CLI returned empty token" });
+        }
       } else {
-        res.statusCode = result.status;
-        res.setHeader("Content-Type", "application/json");
-        res.end(JSON.stringify({ error: result.error }));
+        respond(401, { error: "Failed to refresh guardian token" });
       }
+    });
+
+    child.on("error", (err) => {
+      respond(500, { error: `Failed to spawn CLI: ${err.message}` });
     });
   };
 }
 
+// ---------------------------------------------------------------------------
+// Gateway proxy middleware
+// ---------------------------------------------------------------------------
+
+const GATEWAY_PATTERN = /^(?:\/assistant)?\/__gateway\/(\d+)(\/.*)?$/;
+
+function addPortFromUrl(url: unknown, ports: Set<number>): void {
+  if (typeof url !== "string") return;
+  try {
+    const parsed = new URL(url);
+    if (parsed.hostname !== "127.0.0.1" && parsed.hostname !== "localhost") return;
+    const port = Number(parsed.port);
+    if (Number.isInteger(port) && port >= 1024 && port <= 65535) {
+      ports.add(port);
+    }
+  } catch {
+    // malformed URL — skip
+  }
+}
+
+function readAllowedGatewayPorts(lockfilePaths: string[]): Set<number> {
+  const ports = new Set<number>();
+  for (const candidate of lockfilePaths) {
+    try {
+      const raw = fs.readFileSync(candidate, "utf-8");
+      const data = JSON.parse(raw) as {
+        assistants?: Array<{
+          gatewayUrl?: unknown;
+          localUrl?: unknown;
+          resources?: { gatewayPort?: unknown };
+        }>;
+      };
+      const assistants = Array.isArray(data.assistants) ? data.assistants : [];
+      for (const assistant of assistants) {
+        if (!assistant) continue;
+        addPortFromUrl(assistant.gatewayUrl, ports);
+        addPortFromUrl(assistant.localUrl, ports);
+        const gp = assistant.resources?.gatewayPort;
+        if (typeof gp === "number" && Number.isInteger(gp) && gp >= 1024 && gp <= 65535) {
+          ports.add(gp);
+        }
+      }
+      if (ports.size > 0) return ports;
+    } catch (err: unknown) {
+      if ((err as NodeJS.ErrnoException).code !== "ENOENT") return new Set<number>();
+    }
+  }
+  return ports;
+}
+
+/**
+ * Connect middleware that proxies requests to local gateway ports.
+ *
+ * Matches `/__gateway/:port/*` and forwards to `http://127.0.0.1:{port}{rest}`.
+ * Supports chunked transfer and SSE by piping without buffering.
+ */
 function gatewayProxyMiddleware(lockfilePaths: string[]): Connect.NextHandleFunction {
   return (req, res, next) => {
-    const result = parseGatewayUrl(req.url ?? "");
-    if (!result.match) return next();
+    const match = req.url?.match(GATEWAY_PATTERN);
+    if (!match) return next();
 
     if (rejectUnlessLoopback(req, res)) return;
 
-    if (!result.valid) {
+    const port = parseInt(match[1]!, 10);
+    if (port < 1024 || port > 65535) {
       res.statusCode = 400;
       res.end("Port must be between 1024 and 65535");
       return;
     }
 
-    const { target } = result;
     const allowedPorts = readAllowedGatewayPorts(lockfilePaths);
-    if (!allowedPorts.has(target.port)) {
+    if (!allowedPorts.has(port)) {
       res.statusCode = 403;
       res.end("Gateway port is not active in lockfile");
       return;
     }
 
+    const restPath = match[2] || "/";
+
     const proxyOptions: http.RequestOptions = {
       hostname: "127.0.0.1",
-      port: target.port,
-      path: target.path,
+      port,
+      path: restPath,
       method: req.method,
-      headers: { ...req.headers, host: `127.0.0.1:${target.port}` },
+      headers: { ...req.headers, host: `127.0.0.1:${port}` },
     };
 
     const proxyReq = http.request(proxyOptions, (proxyRes) => {
@@ -318,6 +893,7 @@ function gatewayProxyMiddleware(lockfilePaths: string[]): Connect.NextHandleFunc
       }
     });
 
+    // Pipe the incoming request body to the proxy request (supports POST, etc.)
     req.pipe(proxyReq);
   };
 }
