--- cli/src/lib/local-endpoints/index.ts
@@ -1,13 +1,0 @@
-export { stripSensitiveFields, isLoopbackAddr, resolveCliPath, resetCliPathCache } from "./util";
-export { resolveLocalConfigFromEnv, resolveLockfilePaths, resolveConfigDir } from "./config";
-export type { LocalEndpointConfig } from "./config";
-export { getLockfileData, upsertLockfileAssistant } from "./lockfile";
-export type { LockfileResult, WriteResult } from "./lockfile";
-export { runHatch } from "./hatch";
-export type { HatchResult } from "./hatch";
-export { runRetire } from "./retire";
-export type { RetireResult } from "./retire";
-export { getGuardianAccessToken } from "./guardian-token";
-export type { TokenResult } from "./guardian-token";
-export { parseGatewayUrl, readAllowedGatewayPorts } from "./gateway-proxy";
-export type { GatewayTarget, GatewayParseResult } from "./gateway-proxy";
